requirements.txt
requests
colorama
pycryptodome